Change WatchFunctions into a function.
